agricyle_backend/requirements.txt
Django>=5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
mysqlclient>=2.2
openpyxl>=3.1
reportlab>=4.0
pillow>=10.0
python-dotenv>=1.0
twilio>=9.0
resend
flask==2.3.3
flask-cors==4.0.0
flask-mysql-connector==1.0.0
python-dotenv==1.0.0
mysql-connector-python==8.1.0